+++ pkg/prometheus/promcfg_test.go
@@ -4324,6 +4324,26 @@ func TestRemoteWriteConfig(t *testing.T) {
 			},
 			golden: "RemoteWriteConfig_v3.1.0.golden",
 		},
+		{
+			version: "v2.28.0",
+			remoteWrite: monitoringv1.RemoteWriteSpec{
+				URL: "http://example.com",
+				MetadataConfig: &monitoringv1.MetadataConfig{
+					MaxSamplesPerSend: ptr.To(int32(10)),
+				},
+			},
+			golden: "RemoteWriteConfig_v2.28.0_MaxSamplesPerSendMetadataConfig.golden",
+		},
+		{
+			version: "v2.29.0",
+			remoteWrite: monitoringv1.RemoteWriteSpec{
+				URL: "http://example.com",
+				MetadataConfig: &monitoringv1.MetadataConfig{
+					MaxSamplesPerSend: ptr.To(int32(10)),
+				},
+			},
+			golden: "RemoteWriteConfig_v2.29.0_MaxSamplesPerSendMetadataConfig.golden",
+		},
 	} {
 		t.Run(fmt.Sprintf("i=%d,version=%s", i, tc.version), func(t *testing.T) {
 			p := defaultPrometheus()
@@ -4840,13 +4860,24 @@ func TestNativeHistogramConfig(t *testing.T) {
 		golden                string
 	}{
 		{
+			version: "v3.0.0",
+			nativeHistogramConfig: monitoringv1.NativeHistogramConfig{
+				NativeHistogramBucketLimit:     ptr.To(uint64(10)),
+				ScrapeClassicHistograms:        ptr.To(true),
+				NativeHistogramMinBucketFactor: ptr.To(resource.MustParse("12.124")),
+				ConvertClassicHistogramsToNHCB: ptr.To(true),
+			},
+			golden: "NativeHistogramConfig.golden",
+		},
+		{
 			version: "v2.54.0",
 			nativeHistogramConfig: monitoringv1.NativeHistogramConfig{
 				NativeHistogramBucketLimit:     ptr.To(uint64(10)),
 				ScrapeClassicHistograms:        ptr.To(true),
 				NativeHistogramMinBucketFactor: ptr.To(resource.MustParse("12.124")),
+				ConvertClassicHistogramsToNHCB: ptr.To(true),
 			},
-			golden: "NativeHistogramConfig.golden",
+			golden: "NativeHistogramConfigMissConvertClassicHistogramsToNHCB.golden",
 		},
 		{
 			version: "v2.46.0",
@@ -4854,6 +4885,7 @@ func TestNativeHistogramConfig(t *testing.T) {
 				NativeHistogramBucketLimit:     ptr.To(uint64(10)),
 				ScrapeClassicHistograms:        ptr.To(true),
 				NativeHistogramMinBucketFactor: ptr.To(resource.MustParse("12.124")),
+				ConvertClassicHistogramsToNHCB: ptr.To(true),
 			},
 			golden: "NativeHistogramConfigWithMissNativeHistogramMinBucketFactor.golden",
 		},
@@ -4863,6 +4895,7 @@ func TestNativeHistogramConfig(t *testing.T) {
 				NativeHistogramBucketLimit:     ptr.To(uint64(10)),
 				ScrapeClassicHistograms:        ptr.To(true),
 				NativeHistogramMinBucketFactor: ptr.To(resource.MustParse("12.124")),
+				ConvertClassicHistogramsToNHCB: ptr.To(true),
 			},
 			golden: "NativeHistogramConfigWithMissALL.golden",
 		},
@@ -4872,6 +4905,7 @@ func TestNativeHistogramConfig(t *testing.T) {
 				NativeHistogramBucketLimit:     ptr.To(uint64(10)),
 				ScrapeClassicHistograms:        ptr.To(true),
 				NativeHistogramMinBucketFactor: ptr.To(resource.MustParse("12.124")),
+				ConvertClassicHistogramsToNHCB: ptr.To(true),
 			},
 			golden: "NativeHistogramConfigAlwaysScrapeClassicHistograms.golden",
 		},
@@ -5930,8 +5964,8 @@ func TestScrapeConfigSpecConfig(t *testing.T) {
 						Action:       "hashmod",
 					},
 					{
-						SourceLabels: []monitoringv1.LabelName{"__tmp_hash"},
-						Regex:        "$(SHARD)",
+						SourceLabels: []monitoringv1.LabelName{"__tmp_hash", "__tmp_disable_sharding"},
+						Regex:        "$(SHARD);|.+;.+",
 						Action:       "keep",
 					},
 				},
@@ -6555,6 +6589,70 @@ func TestScrapeConfigSpecConfig(t *testing.T) {
 				},
 			},
 			golden: "ScrapeConfigSpecConfig_WithOAuth_Unsupported.golden",
+		},
+		{
+			name:    "config_NameValidationScheme_utf8",
+			version: "v3.0.0",
+			scSpec: monitoringv1alpha1.ScrapeConfigSpec{
+				NameValidationScheme: ptr.To(monitoringv1.UTF8NameValidationScheme),
+			},
+			golden: "NameValidationScheme_utf8.golden",
+		},
+		{
+			name:    "config_NameValidationScheme_legacy",
+			version: "v3.0.0",
+			scSpec: monitoringv1alpha1.ScrapeConfigSpec{
+				NameValidationScheme: ptr.To(monitoringv1.LegacyNameValidationScheme),
+			},
+			golden: "NameValidationScheme_legacy.golden",
+		},
+		{
+			name:    "config_NameValidationScheme_unsupported",
+			version: "v2.55.0",
+			scSpec: monitoringv1alpha1.ScrapeConfigSpec{
+				NameValidationScheme: ptr.To(monitoringv1.UTF8NameValidationScheme),
+			},
+			golden: "NameValidationScheme_unsupported.golden",
+		},
+		{
+			name:    "config_NameEscapingScheme_AllowUTF8",
+			version: "v3.4.0",
+			scSpec: monitoringv1alpha1.ScrapeConfigSpec{
+				NameEscapingScheme: ptr.To(monitoringv1.AllowUTF8NameEscapingScheme),
+			},
+			golden: "NameEscapingScheme_AllowUTF8.golden",
+		},
+		{
+			name:    "config_NameEscapingScheme_Underscores",
+			version: "v3.4.0",
+			scSpec: monitoringv1alpha1.ScrapeConfigSpec{
+				NameEscapingScheme: ptr.To(monitoringv1.UnderscoresNameEscapingScheme),
+			},
+			golden: "NameEscapingScheme_Underscores.golden",
+		},
+		{
+			name:    "config_NameEscapingScheme_Dots",
+			version: "v3.4.0",
+			scSpec: monitoringv1alpha1.ScrapeConfigSpec{
+				NameEscapingScheme: ptr.To(monitoringv1.DotsNameEscapingScheme),
+			},
+			golden: "NameEscapingScheme_Dots.golden",
+		},
+		{
+			name:    "config_NameEscapingScheme_Values",
+			version: "v3.4.0",
+			scSpec: monitoringv1alpha1.ScrapeConfigSpec{
+				NameEscapingScheme: ptr.To(monitoringv1.ValuesNameEscapingScheme),
+			},
+			golden: "NameEscapingScheme_Values.golden",
+		},
+		{
+			name:    "config_NameEscapingScheme_Unsupported",
+			version: "v3.3.0",
+			scSpec: monitoringv1alpha1.ScrapeConfigSpec{
+				NameEscapingScheme: ptr.To(monitoringv1.ValuesNameEscapingScheme),
+			},
+			golden: "NameEscapingScheme_Unsupported.golden",
 		},
 	} {
 		t.Run(tc.name, func(t *testing.T) {
@@ -9220,13 +9318,13 @@ func TestAppendNameValidationScheme(t *testing.T) {
 	}{
 		{
 			name:                 "UTF8 nameValidationScheme withPrometheus Version 3",
-			version:              "v3.0.0-beta.0",
+			version:              "v3.0.0",
 			nameValidationScheme: ptr.To(monitoringv1.UTF8NameValidationScheme),
 			expectedCfg:          "NameValidationSchemeUTF8WithPrometheusV3.golden",
 		},
 		{
 			name:                 "Legacy nameValidationScheme with Prometheus Version 3",
-			version:              "v3.0.0-beta.0",
+			version:              "v3.0.0",
 			nameValidationScheme: ptr.To(monitoringv1.LegacyNameValidationScheme),
 			expectedCfg:          "NameValidationSchemeLegacyWithPrometheusV3.golden",
 		},
@@ -9273,6 +9371,121 @@ func TestAppendNameValidationScheme(t *testing.T) {
 		})
 	}
 }
+
+func TestAppendNameEscapingScheme(t *testing.T) {
+	testCases := []struct {
+		name               string
+		version            string
+		nameEscapingScheme *monitoringv1.NameEscapingSchemeOptions
+		expectedCfg        string
+	}{
+		{
+			name:               "AllowUTF8 nameEscapingScheme withPrometheus Version 3",
+			version:            "v3.4.0",
+			nameEscapingScheme: ptr.To(monitoringv1.AllowUTF8NameEscapingScheme),
+			expectedCfg:        "NameEscapingSchemeUTF8WithPrometheusV3.golden",
+		},
+		{
+			name:               "Underscores nameEscapingScheme with Prometheus Version 3",
+			version:            "v3.4.0",
+			nameEscapingScheme: ptr.To(monitoringv1.UnderscoresNameEscapingScheme),
+			expectedCfg:        "NameEscapingSchemeUnderscoresWithPrometheusV3.golden",
+		},
+		{
+			name:               "Underscores nameEscapingScheme with Prometheus Version 3",
+			version:            "v2.55.0",
+			nameEscapingScheme: ptr.To(monitoringv1.UnderscoresNameEscapingScheme),
+			expectedCfg:        "NameEscapingSchemeUnderscoresWithPrometheusLowerThanV3.golden",
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+
+			p := defaultPrometheus()
+			if tc.version != "" {
+				p.Spec.CommonPrometheusFields.Version = tc.version
+			}
+			if tc.nameEscapingScheme != nil {
+				p.Spec.CommonPrometheusFields.NameEscapingScheme = tc.nameEscapingScheme
+			}
+
+			cg := mustNewConfigGenerator(t, p)
+			cfg, err := cg.GenerateServerConfiguration(
+				p,
+				nil,
+				nil,
+				nil,
+				nil,
+				&assets.StoreBuilder{},
+				nil,
+				nil,
+				nil,
+				nil,
+			)
+			require.NoError(t, err)
+
+			golden.Assert(t, string(cfg), tc.expectedCfg)
+		})
+	}
+}
+
+func TestAppendConvertClassicHistogramsToNHCB(t *testing.T) {
+	testCases := []struct {
+		name                           string
+		version                        string
+		convertClassicHistogramsToNHCB *bool
+		expectedCfg                    string
+	}{
+		{
+			name:                           "ConvertClassicHistogramsToNHCB true with Prometheus Version 3.4",
+			version:                        "v3.4.0",
+			convertClassicHistogramsToNHCB: ptr.To(true),
+			expectedCfg:                    "ConvertClassicHistogramsToNHCBTrueWithPrometheusV3.golden",
+		},
+		{
+			name:                           "ConvertClassicHistogramsToNHCB false with Prometheus Version 3.4",
+			version:                        "v3.4.0",
+			convertClassicHistogramsToNHCB: ptr.To(false),
+			expectedCfg:                    "ConvertClassicHistogramsToNHCBFalseWithPrometheusV3.golden",
+		},
+		{
+			name:                           "ConvertClassicHistogramsToNHCB true with Prometheus Version 2",
+			version:                        "v2.55.0",
+			convertClassicHistogramsToNHCB: ptr.To(true),
+			expectedCfg:                    "ConvertClassicHistogramsToNHCBTrueWithPrometheusLowerThanV3.golden",
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+
+			p := defaultPrometheus()
+			if tc.version != "" {
+				p.Spec.CommonPrometheusFields.Version = tc.version
+			}
+			if tc.convertClassicHistogramsToNHCB != nil {
+				p.Spec.CommonPrometheusFields.ConvertClassicHistogramsToNHCB = tc.convertClassicHistogramsToNHCB
+			}
+
+			cg := mustNewConfigGenerator(t, p)
+			cfg, err := cg.GenerateServerConfiguration(
+				p,
+				nil,
+				nil,
+				nil,
+				nil,
+				&assets.StoreBuilder{},
+				nil,
+				nil,
+				nil,
+				nil,
+			)
+			require.NoError(t, err)
+
+			golden.Assert(t, string(cfg), tc.expectedCfg)
+		})
+	}
+}
+
 func TestOTLPConfig(t *testing.T) {
 	testCases := []struct {
 		otlpConfig    *monitoringv1.OTLPConfig
@@ -9357,12 +9570,44 @@ func TestOTLPConfig(t *testing.T) {
 			golden: "OTLPConfig_Config_keep_identifying_resource_attributes.golden",
 		},
 		{
-			name:    "Config KeepIdentifyingResourceAttributes with old version",
+			name:    "Config ConvertHistogramsToNHCB old version",
 			version: "v3.0.0",
 			otlpConfig: &monitoringv1.OTLPConfig{
 				KeepIdentifyingResourceAttributes: ptr.To(false),
 			},
 			golden: "OTLPConfig_Config_keep_identifying_resource_attributes_with_old_version.golden",
+		},
+		{
+			name:    "Config NoTranslation translation strategy",
+			version: "v3.4.0",
+			otlpConfig: &monitoringv1.OTLPConfig{
+				TranslationStrategy: ptr.To(monitoringv1.NoTranslation),
+			},
+			golden: "OTLPConfig_Config_translation_strategy_with_notranslation.golden",
+		},
+		{
+			name:    "Config NoTranslation translation strategye unsupported version",
+			version: "v3.0.0",
+			otlpConfig: &monitoringv1.OTLPConfig{
+				TranslationStrategy: ptr.To(monitoringv1.NoTranslation),
+			},
+			expectedErr: true,
+		},
+		{
+			name:    "Config ConvertHistogramsToNHCB",
+			version: "v3.4.0",
+			otlpConfig: &monitoringv1.OTLPConfig{
+				ConvertHistogramsToNHCB: ptr.To(true),
+			},
+			golden: "OTLPConfig_Config_convert_histograms_to_nhcb.golden",
+		},
+		{
+			name:    "Config ConvertHistogramsToNHCB with old version",
+			version: "v3.3.1",
+			otlpConfig: &monitoringv1.OTLPConfig{
+				ConvertHistogramsToNHCB: ptr.To(true),
+			},
+			golden: "OTLPConfig_Config_convert_histograms_to_nhcb_with_old_version.golden",
 		},
 	}
 	for _, tc := range testCases {